+++ src/Git_Updater_PRO/REST/REST_API.php
@@ -83,13 +83,14 @@ class REST_API {
 			self::$namespace,
 			'plugins-api',
 			[
-				'show_in_index'       => false,
+				'show_in_index'       => true,
 				'methods'             => \WP_REST_Server::READABLE,
 				'callback'            => [ $this, 'get_plugins_api_data' ],
 				'permission_callback' => '__return_true',
 				'args'                => [
-					'repo' => [
+					'slug' => [
 						'default'           => false,
+						'required'          => true,
 						'validate_callback' => 'sanitize_text_field',
 					],
 				],
@@ -303,46 +304,49 @@ class REST_API {
 	/**
 	 * Get specific repo plugin API data.
 	 *
+	 * Returns data consistent with `plugins_api()` request.
+	 *
 	 * @param \WP_REST_Request $request REST API response.
 	 *
-	 * @return array
+	 * @return array|\WP_Error
 	 */
 	public function get_plugins_api_data( \WP_REST_Request $request ) {
-		$slug             = $request->get_param( 'repo' );
-		$plugins_api_data = [ 'error' => "The plugin slug, {$slug}, does not exist." ];
-		$gu_plugins       = Singleton::get_instance( 'Fragen\Git_Updater\Plugin', $this )->get_plugin_configs();
+		$slug       = $request->get_param( 'slug' );
+		$gu_plugins = Singleton::get_instance( 'Fragen\Git_Updater\Plugin', $this )->get_plugin_configs();
 
-		if ( array_key_exists( $slug, $gu_plugins ) ) {
-			$repo_data = Singleton::get_instance( 'Fragen\Git_Updater\Base', $this )->get_remote_repo_meta( $gu_plugins[ $slug ] );
-
-			if ( ! property_exists( $repo_data, 'slug' ) ) {
-				return $plugins_api_data;
-			}
-
-			$plugins_api_data = [
-				'name'              => $repo_data->name,
-				'slug'              => $repo_data->slug,
-				'type'              => $repo_data->type,
-				'version'           => $repo_data->remote_version,
-				'author'            => $repo_data->author,
-				'contributors'      => $repo_data->contributors,
-				'requires'          => $repo_data->requires,
-				'tested'            => $repo_data->tested,
-				'requires_php'      => $repo_data->requires_php,
-				'sections'          => $repo_data->sections,
-				// phpcs:ignore WordPress.WP.AlternativeFunctions.strip_tags_strip_tags
-				'short_description' => substr( strip_tags( trim( $repo_data->sections['description'] ) ), 0, 175 ) . '...',
-				'primary_branch'    => $repo_data->primary_branch,
-				'branch'            => $repo_data->branch,
-				'download_link'     => $repo_data->download_link,
-				'banners'           => $repo_data->banners,
-				'icons'             => $repo_data->icons,
-				'last_updated'      => $repo_data->last_updated,
-				'num_ratings'       => $repo_data->num_ratings,
-				'rating'            => $repo_data->rating,
-				'active_installs'   => $repo_data->downloaded,
-			];
+		if ( ! \array_key_exists( $slug, $gu_plugins ) ) {
+			return new \WP_Error( 'no_plugin', 'Specified plugin does not exist.' );
 		}
+
+		$repo_data = Singleton::get_instance( 'Fragen\Git_Updater\Base', $this )->get_remote_repo_meta( $gu_plugins[ $slug ] );
+
+		if ( ! property_exists( $repo_data, 'slug' ) ) {
+			return new \WP_Error( 'bad_data', 'Plugin data is incomplete.' );
+		}
+
+		$plugins_api_data = [
+			'name'              => $repo_data->name,
+			'slug'              => $repo_data->slug,
+			'type'              => $repo_data->type,
+			'version'           => $repo_data->remote_version,
+			'author'            => $repo_data->author,
+			'contributors'      => $repo_data->contributors,
+			'requires'          => $repo_data->requires,
+			'tested'            => $repo_data->tested,
+			'requires_php'      => $repo_data->requires_php,
+			'sections'          => $repo_data->sections,
+			// phpcs:ignore WordPress.WP.AlternativeFunctions.strip_tags_strip_tags
+			'short_description' => substr( strip_tags( trim( $repo_data->section['description'] ) ), 0, 175 ) . '...',
+			'primary_branch'    => $repo_data->primary_branch,
+			'branch'            => $repo_data->branch,
+			'download_link'     => $repo_data->download_link,
+			'banners'           => $repo_data->banners,
+			'icons'             => $repo_data->icons,
+			'last_updated'      => $repo_data->last_updated,
+			'num_ratings'       => $repo_data->num_ratings,
+			'rating'            => $repo_data->rating,
+			'active_installs'   => $repo_data->downloaded,
+		];
 
 		return $plugins_api_data;
 	}